Make resolveAllOf function a new service
